% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_yes_no.R
\name{ask_yes_no}
\alias{ask_yes_no}
\title{Ask yes no}
\usage{
ask_yes_no(prompt, boolean = TRUE)
}
\arguments{
\item{prompt}{character, prompt to ask the user.}

\item{boolean}{boolean, whether to return results as boolean or as 'y/n'.}
}
\description{
This function allows yes/no answers to be returned as TRUE/FALSE or 'y'/'n'. It will prompt user multiple times until a valid response is given.
}
\seealso{
`utils::askYesNo()`
}
