% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_methods.R
\name{available_methods}
\alias{available_methods}
\title{Available methods}
\usage{
available_methods(quietly = TRUE)
}
\arguments{
\item{quietly}{Boolean, allows user to return a `data.frame` with available methods.}
}
\description{
This function looks for available methods and creates global objects required by other functions in the package.
}
