% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_directory.R
\name{choose_directory}
\alias{choose_directory}
\title{Choose directory}
\usage{
choose_directory(initial_dir = getwd(),
  method = select_directory_method(), title = NULL)
}
\arguments{
\item{initial_dir}{Path to folder from which to display dialog. Default is to use current working directory.}

\item{method}{Package to be used to construct the GUI. See `select_directory_method()`}

\item{title}{Character, title of the vector}
}
\description{
This function allows the user to choose a directory interactively. It has been tested on Windows, Ubuntu and macOS. The gist comes from https://stackoverflow.com/questions/48218491/os-independent-way-to-select-directory-interactively-in-r
}
\seealso{
`select_directory_method()`
}
