% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_files.R
\name{choose_files}
\alias{choose_files}
\title{Choose files}
\usage{
choose_files(initial_dir = getwd(), method = select_directory_method(),
  title = NULL, multiple = TRUE)
}
\arguments{
\item{initial_dir}{Path to folder from which to display dialog. Default is to use current working directory.}

\item{method}{Package to be used to construct the GUI. See `select_directory_method()`}

\item{title}{Character, title of the vector}

\item{multiple}{Boolean, allows for multiple files to be selected. Some GUIs do not allow multiple files.}
}
\description{
This function allows the user to choose file(s) interactively. It has been tested on Windows, Ubuntu and macOS.
}
\seealso{
select_directory_method(), choose_directory()
}
