% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_menu.R
\name{numeric_menu}
\alias{numeric_menu}
\title{Numeric menu}
\usage{
numeric_menu(c("apple", "banana", "coconut"), "What would you like for dessert?")
}
\arguments{
\item{opts}{vector with options to be displayed on the menu}

\item{prompt}{character vector with the propmt (default to "Use numbers to select the option").}
}
\description{
This function allows the user to create numeric menus with ease for interaction with user. The menu is robust to selecting orders that are not available or are not numbers.
}
